backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-aiplatform
google-generativeai
genkit[googleai]
firebase-admin
pydantic
python-docx
pdfplumber
weasyprint
google-api-python-client
sendgrid
